app/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]

